refactor(authentication): 2fa handler implementation (#352)
