+++ modules/authentication/src/constants/TokenType.ts
@@ -1,8 +1,8 @@
 export enum TokenType {
   VERIFICATION_TOKEN = 'VERIFICATION_TOKEN',
   PASSWORD_RESET_TOKEN = 'PASSWORD_RESET_TOKEN',
-  CHANGE_PASSWORD_TOKEN = 'CHANGE_PASSWORD_TOKEN',
   CHANGE_EMAIL_TOKEN = 'CHANGE_EMAIL_TOKEN',
+  TWO_FA_CHANGE_PASSWORD_TOKEN = 'TWO_FA_CHANGE_PASSWORD_TOKEN',
   PHONE_TWO_FA_VERIFICATION_TOKEN = 'PHONE_TWO_FA_VERIFICATION_TOKEN',
   QR_TWO_FA_VERIFICATION_TOKEN = 'QR_TWO_FA_VERIFICATION_TOKEN',
   VERIFY_PHONE_NUMBER_TOKEN = 'VERIFY_PHONE_NUMBER_TOKEN',
